calculation total sum of refunding products

+++ data/src/main/java/kg/nurtelecom/data/history/Content.kt
@@ -1,7 +1,9 @@
 package kg.nurtelecom.data.history
 
+import java.math.BigDecimal
+
 data class Content (
-	val total : Double,
+	val total : BigDecimal,
 	val gnsRegNum : Int,
 	val indexNum : Int,
 	val subtotal : Int,
